requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
APScheduler==3.10.4
plyer==2.0.0